+++ pages/base_page.py
@@ -1,0 +1,8 @@
+class BasePage():
+
+    def __init__(self, browser, url):
+        self.browser = browser
+        self.url = url
+
+    def open(self):
+        self.browser.get(self.url)
