+++ pages/main_page.py
@@ -1,8 +1,2 @@
-class BasePage():
+from .base_page import BasePage
 
-    def __init__(self, browser, url):
-        self.browser = browser
-        self.url = url
-
-    def open(self):
-        self.browser.get(self.url)
